Only generate pre-release for merges with changes on src or types
create tag with the beta version
